+++ app/api/__init__.py
@@ -1,0 +1,4 @@
+"""Router for /api endpoints."""
+from app.api.router import router
+
+__all__ = ["router"]
