+++ app/api/predict/__init__.py
@@ -1,0 +1,4 @@
+"""Router for predict tools in api endpoint."""
+from app.api.predict.router import router
+
+__all__ = ["router"]
