+++ app/web/generate/__init__.py
@@ -1,0 +1,4 @@
+"""Package with router for generating."""
+from app.web.generate.router import router
+
+__all__ = ["router"]
