+++ docker-compose.yml
@@ -20,11 +20,11 @@ services:
     volumes:
       - .aws/:/root/.aws:ro
 
-  bot:
-    build: ./bot
-    restart: always
-    depends_on:
-      - app
+  # bot:
+  #   build: ./bot
+  #   restart: always
+  #   depends_on:
+  #     - app
 
   redis:
     image: "redis:latest"
